further model refactor
