+++ src/carapace/notifications/store.py
@@ -8,7 +8,7 @@ from threading import RLock
 
 import yaml
 
-from carapace.models.notifications import NotificationPreferences, NotificationSubscription
+from carapace.notifications.models import NotificationPreferences, NotificationSubscription
 
 
 def derive_owner_key(token: str) -> str:
